+++ chaosaws/__init__.py
@@ -240,5 +240,6 @@ def load_exported_activities() -> List[DiscoveredActivities]:
     activities.extend(discover_probes("chaosaws.rds.probes"))
     activities.extend(discover_actions("chaosaws.elasticache.actions"))
     activities.extend(discover_probes('chaosaws.elasticache.probes'))
-
+    activities.extend(discover_actions("chaosaws.emr.actions"))
+    activities.extend(discover_probes('chaosaws.emr.probes'))
     return activities