+++ chaosaws/ecs/actions.py
@@ -1,9 +1,10 @@
 # -*- coding: utf-8 -*-
 import random
 import re
-from typing import List, Union
+from typing import List, Union, Dict
 
 import boto3
+from botocore.exceptions import ClientError
 from chaoslib.exceptions import FailedActivity
 from chaoslib.types import Configuration, Secrets
 from logzero import logger
@@ -11,57 +12,53 @@ from logzero import logger
 from chaosaws import aws_client
 from chaosaws.types import AWSResponse
 
-__all__ = ["stop_random_tasks", "stop_task", "delete_service",
-           "delete_cluster", "deregister_container_instance",
-           "update_desired_count"]
+__all__ = [
+    "stop_random_tasks", "stop_task", "delete_service", "tag_resource",
+    "delete_cluster", "deregister_container_instance", "untag_resource",
+    "update_desired_count", "set_service_placement_strategy",
+    "set_service_deployment_configuration", "update_container_instances_state"
+]
 
 
-def stop_random_tasks(cluster: str = None,
+def stop_random_tasks(cluster: str,
                       task_count: int = None,
                       task_percent: int = None,
                       service: str = None,
                       reason: str = 'Chaos Testing',
                       configuration: Configuration = None,
-                      secrets: Secrets = None) -> AWSResponse:
+                      secrets: Secrets = None) -> List[AWSResponse]:
     """
     Stop a random number of tasks based on given task_count or task_percent
 
     You can specify a cluster by its ARN identifier or, if not provided, the
     default cluster will be picked up.
 
-    Parameters:
-    Required:
-        - cluster: name of the cluster to stop tasks in
-
-    Optional:
-        - service: name of the service to stop tasks in
-
-    One Of:
-        - task_count: the number of tasks to stop
-        - task_percent: the percentage of tasks to stop
+    :param cluster: The ECS cluster Name
+    :param task_count: The number of tasks to stop
+    :param task_percent: The percentage of total tasks to stop
+    :param service: The ECS service name
+    :param reason: An explanation of why the service was stopped
+    :param configuration: access values used by actions/probes
+    :param secrets: values that need to be passed on to actions/probes
+    :return: List[Dict[str, Any]]
     """
-    client = aws_client("ecs", configuration, secrets)
-
-    if not cluster:
-        raise FailedActivity("A cluster name is required")
-
     if not any([task_count, task_percent]) or all([task_count, task_percent]):
         raise FailedActivity(
             'Must specify one of "task_count", "task_percent"')
 
-    tasks = list_running_tasks_in_cluster(cluster=cluster, client=client,
-                                          service=service)
+    client = aws_client("ecs", configuration, secrets)
+    validate(client, cluster, service)
+
+    tasks = list_running_tasks_in_cluster(
+        cluster=cluster, client=client, service=service)
 
     if task_percent:
-        task_count = int(float(
-            len(tasks) * float(task_percent)) / 100)
+        task_count = int(float(len(tasks) * float(task_percent)) / 100)
 
     if len(tasks) < task_count:
         raise FailedActivity(
             'Not enough running tasks in {} to satisfy '
-            'stop count {} ({})'.format(
-                cluster, task_count,
-                len(tasks)))
+            'stop count {} ({})'.format(cluster, task_count, len(tasks)))
 
     tasks = random.sample(tasks, task_count)
 
@@ -83,7 +80,6 @@ def stop_task(cluster: str = None, task_id: str = None, service: str = None,
     """
     Stop a given ECS task instance. If no task_id provided, a random task of
     the given service is stopped.
-
     You can specify a cluster by its ARN identifier or, if not provided, the
     default cluster will be picked up.
     """
@@ -149,7 +145,12 @@ def delete_cluster(cluster: str,
                    configuration: Configuration = None,
                    secrets: Secrets = None) -> AWSResponse:
     """
-    Delete a given ECS cluster
+    Delete an ECS cluster
+
+    :param cluster: The ECS cluster name or ARN
+    :param configuration: access values used by actions/probes
+    :param secrets: values that need to be passed on to actions/probes
+    :return: Dict[str, Any]
     """
     client = aws_client("ecs", configuration, secrets)
     logger.debug("Deleting ECS cluster: {}".format(cluster))
@@ -162,16 +163,24 @@ def deregister_container_instance(cluster: str,
                                   configuration: Configuration = None,
                                   secrets: Secrets = None) -> AWSResponse:
     """
-    Deregister a given ECS container. Becareful that tasks handled by this
-    instance will remain orphan.
+    Deregister an ECS container
+
+    Warning: If using "force", Any tasks not deleted before deregistration
+    will remain orphaned
+
+    :param cluster: The ECS cluster name or ARN or ARN
+    :param instance_id: The container instance id or ARN
+    :param force: Force deregistraion of container instance
+    :param configuration: access values used by actions/probes
+    :param secrets: values that need to be passed on to actions/probes
+    :return: Dict[str, Any]
     """
     client = aws_client("ecs", configuration, secrets)
     logger.debug(
         "Deregistering container instance '{}' from ECS cluster: {}".format(
             instance_id, cluster))
-    return client.deregister_container_instance(cluster=cluster,
-                                                containerInstance=instance_id,
-                                                force=force)
+    return client.deregister_container_instance(
+        cluster=cluster, containerInstance=instance_id, force=force)
 
 
 def update_desired_count(cluster: str,
@@ -179,9 +188,17 @@ def update_desired_count(cluster: str,
                          desired_count: int,
                          configuration: Configuration = None,
                          secrets: Secrets = None) -> AWSResponse:
-    """Allows for changing the desired task count value for a given ecs service
+    """
+    Set the number of desired tasks for an ECS service
 
-    Action Example:
+    :param cluster: The ECS cluster name or ARN or ARN
+    :param service: The ECS service name
+    :param desired_count: The number of instantiation of the tasks to run
+    :param configuration: access values used by actions/probes
+    :param secrets: values that need to be passed on to actions/probes
+    :return: Dict[str, Any]
+
+    Example:
         "method": {
             "type": "action",
             "name": "update service",
@@ -198,6 +215,38 @@ def update_desired_count(cluster: str,
         }
     """
     client = aws_client("ecs", configuration, secrets)
+    validate(client, cluster, service)
+
+    return client.update_service(
+        cluster=cluster, service=service, desiredCount=desired_count)
+
+
+def set_service_deployment_configuration(
+        cluster: str,
+        service: str,
+        maximum_percent: int = 200,
+        minimum_healthy_percent: int = 100,
+        configuration: Configuration = None,
+        secrets: Secrets = None) -> AWSResponse:
+    """
+    Sets the maximum healthy count and minimum healthy percentage values for
+    a services deployment configuration
+
+    :param cluster: The ECS cluster name or ARN
+    :param service: The ECS service name
+    :param maximum_percent: The upper limit on the number of tasks a service is
+        allowed to have in RUNNING or PENDING during deployment
+    :param minimum_healthy_percent: The lower limit on the number of tasks a
+        service must keep in RUNNING to be considered healthy during deployment
+    :param configuration: access values used by actions/probes
+    :param secrets: values that need to be passed on to actions/probes
+    :return: Dict[str, Any]
+    """
+    if minimum_healthy_percent > maximum_percent:
+        raise FailedActivity('minimum_healthy_percent cannot be larger '
+                             'than maximum_percent')
+
+    client = aws_client("ecs", configuration, secrets)
 
     if not validate_cluster(cluster, client):
         raise FailedActivity('unable to locate cluster: %s' % cluster)
@@ -205,15 +254,178 @@ def update_desired_count(cluster: str,
         raise FailedActivity('unable to locate service: %s on %s' % (
             service, cluster))
 
-    return client.update_service(
-        cluster=cluster,
-        service=service,
-        desiredCount=desired_count)
+    params = {
+        'cluster': cluster,
+        'service': service,
+        'deploymentConfiguration': {
+            'maximumPercent': maximum_percent,
+            'minimumHealthyPercent': minimum_healthy_percent
+        }
+    }
+    return client.update_service(**params)
+
+
+def set_service_placement_strategy(cluster: str,
+                                   service: str,
+                                   placement_type: str,
+                                   placement_field: str = None,
+                                   configuration: Configuration = None,
+                                   secrets: Secrets = None) -> AWSResponse:
+    """
+    Sets the service's instance placement strategy
+
+    :param cluster: The ECS cluster name or ARN
+    :param service: The ECS service name
+    :param placement_type: The type of placement strategy to employ
+        (random, spread, or binpack)
+    :param placement_field: The field to apply the strategy against
+        (eg: "attribute:ecs.availability-zone")
+    :param configuration: access values used by actions/probes
+    :param secrets: values that need to be passed on to actions/probes
+    :return: Dict[str, Any]
+    """
+    if placement_type in ('spread', 'binpack') and not placement_field:
+        raise FailedActivity('"placement_field" is required when using '
+                             '"spread" or "binpack"')
+
+    client = aws_client("ecs", configuration, secrets)
+    validate(client, cluster, service)
+
+    if placement_type == 'random':
+        placement_field = None
+
+    params = {
+        'cluster': cluster,
+        'service': service,
+        'placementStrategy': [{
+            'type': placement_type,
+            **({'field': placement_field} if placement_field else {})
+        }]
+    }
+
+    try:
+        return client.update_service(**params)
+    except ClientError as e:
+        logger.exception(e.response['Error']['Message'])
+        raise FailedActivity(e.response['Error']['Message'])
+
+
+def tag_resource(tags: List[Dict[str, str]],
+                 resource_arn: str,
+                 configuration: Configuration = None,
+                 secrets: Secrets = None):
+    """
+    Tags the provided resource(s) with provided tags
+
+    ** For ECS resources, the long form ARN must be used
+    https://docs.aws.amazon.com/AmazonECS/latest/developerguide/ecs-account-settings.html#ecs-resource-arn-timeline
+
+    Example:
+        {
+            "tags": [
+                {"key": "MyTagKey", "value": "MyTagValue"},
+                {"key": "MyOtherTagKey", "value": "MyOtherTagValue"}
+            ],
+            "resource_arn": "arn:aws:ecs:us-east-1:123456789012:cluster/name"
+        }
+
+    :param tags: A list of key/value pairs
+    :param resource_arn: The ARN of the resource to tag.
+        Valid resources: capacity providers, tasks, services, task definitions,
+        clusters, and container instances
+    :param configuration: access values used by actions/probes
+    :param secrets: values that need to be passed on to actions/probes
+    :return: Dict[str, Any]
+    """
+    client = aws_client('ecs', configuration, secrets)
+    try:
+        client.tag_resource(resourceArn=resource_arn, tags=tags)
+    except ClientError as e:
+        logger.exception(e.response['Error']['Message'])
+        raise FailedActivity(e.response['Error']['Message'])
+
+
+def untag_resource(tag_keys: List[str],
+                   resource_arn: str,
+                   configuration: Configuration = None,
+                   secrets: Secrets = None):
+    """
+    Removes the given tags from the provided resource
+
+    ** For ECS resources, the long form ARN must be used
+    https://docs.aws.amazon.com/AmazonECS/latest/developerguide/ecs-account-settings.html#ecs-resource-arn-timeline
+
+    Example:
+        {
+            "tag_keys": ["MyTagKey", "MyOtherTagKey"],
+            "resource_arn": "arn:aws:ecs:...:service/cluster-name/service-name"
+        }
+
+    :param tag_keys: A list of tag keys to remove
+    :param resource_arn: The ARN of the resource to tag.
+        Valid resources: capacity providers, tasks, services, task definitions,
+        clusters, and container instances
+    :param configuration: access values used by actions/probes
+    :param secrets: values that need to be passed on to actions/probes
+    :return: Dict[str, Any]
+    """
+    client = aws_client('ecs', configuration, secrets)
+    try:
+        client.untag_resource(resourceArn=resource_arn, tagKeys=tag_keys)
+    except ClientError as e:
+        logger.exception(e.response['Error']['Message'])
+        raise FailedActivity(e.response['Error']['Message'])
+
+
+def update_container_instances_state(cluster: str,
+                                     container_instances: List[str],
+                                     status: str,
+                                     configuration: Configuration = None,
+                                     secrets: Secrets = None) -> AWSResponse:
+    """
+    Modify the status of an ACTIVE ECS container instance
+
+    :param cluster: The ECS cluster name or ARN
+    :param container_instances: A list of container instance ids for ARNs
+    :param status: The desired instance state (Valid States: ACTIVE, DRAINING)
+    :param configuration: access values used by actions/probes
+    :param secrets: values that need to be passed on to actions/probes
+    :return: Dict[str, Any]
+    """
+    client = aws_client('ecs', configuration, secrets)
+    validate(client, cluster)
+
+    try:
+        return client.update_container_instance_state(
+            cluster=cluster,
+            containerInstances=container_instances,
+            status=status)
+    except ClientError as e:
+        logger.exception(e.response['Error']['Message'])
+        raise FailedActivity(e.response['Error']['Message'])
 
 
 ###############################################################################
 # Private functions
 ###############################################################################
+def validate(client: boto3.client,
+             cluster: str = None,
+             service: str = None):
+    """Validate that a service and/or cluster exists"""
+    if cluster:
+        response = client.describe_clusters(clusters=[cluster])['clusters']
+        if not response:
+            raise FailedActivity('unable to locate cluster: %s' % cluster)
+
+    if service:
+        response = client.describe_services(
+            cluster=cluster, services=[service])['services']
+        if not response:
+            raise FailedActivity(
+                'unable to locate service: %s on cluster: %s' % (
+                    service, cluster))
+
+
 def validate_cluster(cluster: str, client: boto3.client) -> Union[str, None]:
     """Validates the provided cluster exists"""
     cluster = client.describe_clusters(clusters=[cluster])['clusters']
@@ -233,71 +445,68 @@ def validate_service(
 
 
 def list_services_arns(cluster: str, client: boto3.client) -> List[str]:
-    """
-    Return of all services arns in the given cluster.
-    """
-    res = client.list_services(cluster=cluster, maxResults=10)
-    services = res["serviceArns"][:]
-    while True:
-        next_token = res.get("nextToken")
-        if not next_token:
-            break
+    """Return of all services arns in the given cluster."""
+    services = []
 
-        res = client.list_services(
-            cluster=cluster, nextToken=next_token, maxResults=10)
-        services.extend(res["serviceArns"])
+    def _list(next_token=None):
+        params = {
+            'cluster': cluster,
+            **({'nextToken': next_token} if next_token else {})
+        }
+        response = client.list_services(**params)
+        services.extend(response['serviceArns'])
 
+        if response.get('nextToken'):
+            _list(response['nextToken'])
+
+    _list()
     return services
 
 
 def filter_services(services: List[str], pattern: str) -> List[str]:
-    """
-    Return the list of services matching the given pattern.
-    """
+    """Return the list of services matching the given pattern."""
     r = re.compile(pattern)
     return [s for s in services if r.search(s)]
 
 
 def list_tasks(cluster: str, service: str, client: boto3.client) -> List[str]:
-    res = client.list_tasks(cluster=cluster, serviceName=service,
-                            maxResults=100)
-    tasks = res['taskArns'][:]
-    while True:
-        next_token = res.get("nextToken")
-        if not next_token:
-            break
+    tasks = []
 
-        res = client.list_tasks(cluster=cluster, serviceName=service,
-                                nextToken=next_token, maxResults=100)
-        tasks.extend(res["taskArns"])
+    def _list(next_token=None):
+        params = {
+            'cluster': cluster,
+            'maxResults': 100,
+            **({'serviceName': service} if service else {}),
+            **({'nextToken': next_token} if next_token else {})
+        }
+        response = client.list_tasks(**params)
+        tasks.extend(response['taskArns'])
+
+        if response.get('nextToken'):
+            _list(response['nextToken'])
+
+    _list()
     return tasks
 
 
-def list_running_tasks_in_cluster(cluster: str, client: boto3.client,
-                                  service: str = None):
-    if service:
-        res = client.list_tasks(cluster=cluster, serviceName=service,
-                                maxResults=100,
-                                desiredStatus='RUNNING')
-    else:
-        res = client.list_tasks(cluster=cluster,
-                                maxResults=100,
-                                desiredStatus='RUNNING')
-    tasks = res['taskArns'][:]
-    while True:
-        next_token = res.get("nextToken")
-        if not next_token:
-            break
+def list_running_tasks_in_cluster(cluster: str,
+                                  client: boto3.client,
+                                  service: str = None) -> List[str]:
+    tasks = []
 
-        if service:
-            res = client.list_tasks(cluster=cluster, serviceName=service,
-                                    nextToken=next_token,
-                                    maxResults=100,
-                                    desiredStatus='RUNNING')
-        else:
-            res = client.list_tasks(cluster=cluster,
-                                    nextToken=next_token,
-                                    maxResults=100,
-                                    desiredStatus='RUNNING')
-        tasks.extend(res["taskArns"])
+    def _list(next_token=None):
+        params = {
+            'cluster': cluster,
+            'maxResults': 100,
+            'desiredStatus': 'RUNNING',
+            **({'serviceName': service} if service else {}),
+            **({'nextToken': next_token} if next_token else {})
+        }
+        response = client.list_tasks(**params)
+        tasks.extend(response['taskArns'])
+
+        if response.get('nextToken'):
+            _list(response['nextToken'])
+
+    _list()
     return tasks